style deleteBtn and rename files
