+++ client/src/components/PaintingSection/index.css
@@ -63,3 +63,8 @@
     cursor: pointer;
     margin-top: 40%;
   }
+
+  .deleteBtn{
+    margin-left: 200px;
+    margin-top: 20px;
+  }
